2204 patch for DiskSumo
Changes:

- Add R)eceive dump to be able to receive and write a D64 image back to
a floppy disk over Xmodem. For the receiving to work, baud rate has to
be set to 300. Transfer time is roughly 3 hours. A full track of data
is received before writing; during the write the sender may timeout so
you need to adjust Xmodem timeout accordingly (or disable it).

- Fix memory area allocation. Opening the RS-232 device will reset
the string pointer and the end of BASIC area pointer back to default
values. As a result the memory area will eventually get corrupted by
string garbage resulting in bad floppies. Fix by adjusting the pointers
after the device is open.

- Fix communications settings. Open the RS-232 port as documented in
the VIC-1011A manual.

- Fix transfer duration calculation.