2304 patch for DiskSumo
Changes:

- Add s(C)an command for checking the disk for bad sectors.

- Try to fix EOT signaling. Timeouts after 10 seconds/retries.

- Try to optimize for speed and make the BASIC code less verbose.

- Allow to toggle b(I)t rate between 300 and 1200. 1200 bps works reliably
for sending.

- Remove transfer duration reporting. The jiffy counter was too unreliable
during a long transfer.

- Keep the floppy drive open only during the track read/write, so that
drive LED turns off during the track send.

- Implement a counter-based timeout for packet reception to prevent
transfer stalling when characters are dropped mid-packet.

- Allow specifying the S)tart track; this can be used to resume
interrupted transfers. Also the last successfully sent/written track
number is now displayed.

- General notes: The performance is now roughly 100 minutes to send
a floppy dump from C64 (1200 bps) and 185 minutes to receive one
(300 bps). The recommended timeout is 100 seconds on sx side (when
C64 is receiving) and to disable the timeouts on rx side (when C64 is
sending). Note that rx does not recognize EOT properly when the timeout
is disabled, so one needs to end the transfer manually.